+++ mediawiki/mediawikipage.py
@@ -258,7 +258,8 @@ class MediaWikiPage(object):
 
     @property
     def references(self):
-        ''' External links, or references, listed anywhere on the MediaWiki page
+        ''' External links, or references, listed anywhere on the MediaWiki \
+            page
 
         :getter: Returns the list of all external links
         :setter: Not settable
@@ -269,10 +270,8 @@ class MediaWikiPage(object):
         '''
         if self._references is False:
             params = {'prop': 'extlinks', 'ellimit': 'max'}
-            self._references = list()
-            for link in self._continued_query(params):
-                self._references.append(link['*'])
-            self._references = sorted(self._references)
+            tmp = [link['*'] for link in self._continued_query(params)]
+            self._references = sorted(tmp)
         return self._references
 
     @property
@@ -284,18 +283,21 @@ class MediaWikiPage(object):
         :type: list
         '''
         if self._categories is False:
-            self._categories = list()
+
+            def _get_cat(val):
+                ''' parse the category correctly '''
+                tmp = val['title']
+                if tmp.startswith('Category:'):
+                    return tmp[9:]
+                return tmp
+
             params = {
                 'prop': 'categories',
                 'cllimit': 'max',
                 'clshow': '!hidden'
             }
-            for link in self._continued_query(params):
-                cat = link['title']
-                if cat.startswith('Category:'):
-                    cat = cat[9:]
-                self._categories.append(cat)
-            self._categories = sorted(self._categories)
+            tmp = [_get_cat(link) for link in self._continued_query(params)]
+            self._categories = sorted(tmp)
         return self._categories
 
     @property
@@ -338,9 +340,8 @@ class MediaWikiPage(object):
                 'plnamespace': 0,
                 'pllimit': 'max'
             }
-            for link in self._continued_query(params):
-                self._links.append(link['title'])
-            self._links = sorted(self._links)
+            tmp = [link['title'] for link in self._continued_query(params)]
+            self._links = sorted(tmp)
         return self._links
 
     @property
@@ -359,9 +360,8 @@ class MediaWikiPage(object):
                 'rdprop': 'title',
                 'rdlimit': 'max'
             }
-            for link in self._continued_query(params):
-                self._redirects.append(link['title'])
-            self._redirects = sorted(self._redirects)
+            tmp = [link['title'] for link in self._continued_query(params)]
+            self._redirects = sorted(tmp)
         return self._redirects
 
     @property
@@ -382,9 +382,9 @@ class MediaWikiPage(object):
                 'blfilterredir': 'nonredirects',
                 'blnamespace': 0
             }
-            for link in self._continued_query(params, 'backlinks'):
-                self._backlinks.append(link['title'])
-            self._backlinks = sorted(self._backlinks)
+            tmp = [link['title']
+                   for link in self._continued_query(params, 'backlinks')]
+            self._backlinks = sorted(tmp)
         return self._backlinks
 
     @property
@@ -558,14 +558,14 @@ class MediaWikiPage(object):
         lis = BeautifulSoup(html, 'html.parser').find_all('li')
         filtered_lis = [li for li in lis if 'tocsection' not in
                         ''.join(li.get('class', list()))]
-        may_refer_to = [li.a.get_text()
-                        for li in filtered_lis if li.a]
+        may_refer_to = [li.a.get_text() for li in filtered_lis if li.a]
+
         disambiguation = list()
         for lis_item in filtered_lis:
             item = lis_item.find_all('a')
             one_disambiguation = dict()
             one_disambiguation['description'] = lis_item.text
-            if item:
+            if item and hasattr(item, 'title'):
                 one_disambiguation['title'] = item[0]['title']
             else:
                 # these are non-linked records so double up the text
@@ -584,7 +584,6 @@ class MediaWikiPage(object):
                 normalized = query['normalized'][0]
                 if normalized['from'] != self.title:
                     raise MediaWikiException(ODD_ERROR_MESSAGE)
-                # assert normalized['from'] == self.title, ODD_ERROR_MESSAGE
                 from_title = normalized['to']
             else:
                 if not getattr(self, 'title', None):
@@ -593,7 +592,6 @@ class MediaWikiPage(object):
                 from_title = self.title
             if redirects['from'] != from_title:
                 raise MediaWikiException(ODD_ERROR_MESSAGE)
-            # assert redirects['from'] == from_title, ODD_ERROR_MESSAGE
 
             # change the title and reload the whole object
             self.__init__(self.mediawiki, title=redirects['to'],