+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current
 
+### Version 0.3.16
+
+* Add ability to turn off caching completely
+* Fix bug when disambiguation link does not have a title [issue #35](https://github.com/barrust/mediawiki/issues/35)
+
 ### Version 0.3.15
 
 * Add parse all links within a section [issue #33](https://github.com/barrust/mediawiki/issues/33)